requirements.txt
aiohttp==3.7.4.post0
anyio==3.6.1
APScheduler==3.9.1
async-timeout==3.0.1
attrs==21.4.0
cachetools==5.2.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
discord==1.7.3
discord.py==2.0.1
h11==0.12.0
httpcore==0.15.0
httpx==0.23.0
idna==3.3
multidict==6.0.2
python-dotenv==0.19.2
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.2
typing_extensions==4.1.1
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.8
yarl==1.7.2
